Optimise the allocator by using 32-bit offsets
Rather than storing 64-bit pointer to the end of occupied space, store
a 32-bit offset from the start of the heap.  This speeds up alloc by
about a third.  I did not fully investigate how, but two factors that
could have contributed are that a) there’s no longer need to compar
with program input address (only that the 32-bit offset overflows); and
b) 32-bit operations are somehow faster.