Rather than tracking end offset, track used memory
Used memory doesn’t count the header so it may also be thought of as
an end offset measured from the end of the header.  Such formulation
makes it so that value of zero represents properly initialised header
and thus we no longer need to check for zero and perform initialisation.
This removes a branch and speeds up allocation a tad.